+++ lib/reviewer/configuration.rb
@@ -1,19 +1,16 @@
 # frozen_string_literal: true
 
 module Reviewer
-  # Configuration for Reviewer
+  # Configuration values container for Reviewer
   class Configuration
-    DEFAULT_CONFIGURATION_PATH = Dir.pwd.freeze
-    DEFAULT_CONFIGURATION_FILE = '.reviewer.yml'
+    DEFAULT_PATH = Dir.pwd.freeze
+    DEFAULT_FILE_NAME = '.reviewer.yml'
+    DEFAULT_FILE = "#{DEFAULT_PATH}/#{DEFAULT_FILE_NAME}"
 
     attr_accessor :file
 
     def initialize
-      @file = "#{DEFAULT_CONFIGURATION_PATH}/#{DEFAULT_CONFIGURATION_FILE}"
-    end
-
-    def tools
-      @tools ||= Loader.new.to_h
+      @file = DEFAULT_FILE
     end
   end
 end